im-util: Add constructors for Filter and FilterMap
… to allow them to be nested.
