+++ eyeball-im-util/src/vector.rs
@@ -1,12 +1,9 @@
 //! Utilities around [`ObservableVector`].
 
-use std::collections::VecDeque;
-
 use eyeball_im::{ObservableVector, Vector, VectorSubscriber};
 
 mod filter;
 
-use self::filter::FilterImpl;
 pub use self::filter::{Filter, FilterMap};
 
 /// Extension trait for [`ObservableVector`].
@@ -18,7 +15,7 @@ where
     ///
     /// Returns a filtered version of the current vector, and a subscriber to
     /// get updates through.
-    fn subscribe_filter<F>(&self, filter: F) -> (Vector<T>, Filter<VectorSubscriber<T>, F>)
+    fn subscribe_filter<F>(&self, f: F) -> (Vector<T>, Filter<VectorSubscriber<T>, F>)
     where
         F: Fn(&T) -> bool;
 
@@ -27,10 +24,7 @@ where
     ///
     /// Returns a filtered + mapped version of the current vector, and a
     /// subscriber to get updates through.
-    fn subscribe_filter_map<U, F>(
-        &self,
-        filter: F,
-    ) -> (Vector<U>, FilterMap<VectorSubscriber<T>, F>)
+    fn subscribe_filter_map<U, F>(&self, f: F) -> (Vector<U>, FilterMap<VectorSubscriber<T>, F>)
     where
         U: Clone,
         F: Fn(T) -> Option<U>;
@@ -40,52 +34,18 @@ impl<T> VectorExt<T> for ObservableVector<T>
 where
     T: Clone + Send + Sync + 'static,
 {
-    fn subscribe_filter<F>(&self, filter: F) -> (Vector<T>, Filter<VectorSubscriber<T>, F>)
+    fn subscribe_filter<F>(&self, f: F) -> (Vector<T>, Filter<VectorSubscriber<T>, F>)
     where
         F: Fn(&T) -> bool,
     {
-        let mut filtered_indices = VecDeque::new();
-        let mut v = (*self).clone();
-
-        let mut original_idx = 0;
-        v.retain(|val| {
-            let keep = filter(val);
-            if keep {
-                filtered_indices.push_back(original_idx);
-            }
-            original_idx += 1;
-            keep
-        });
-
-        let inner = self.subscribe();
-        let original_len = self.len();
-        let inner = FilterImpl { inner, filtered_indices, original_len };
-        let sub = Filter { inner, filter };
-
-        (v, sub)
+        Filter::new((*self).clone(), self.subscribe(), f)
     }
 
-    fn subscribe_filter_map<U, F>(
-        &self,
-        filter: F,
-    ) -> (Vector<U>, FilterMap<VectorSubscriber<T>, F>)
+    fn subscribe_filter_map<U, F>(&self, f: F) -> (Vector<U>, FilterMap<VectorSubscriber<T>, F>)
     where
         U: Clone,
         F: Fn(T) -> Option<U>,
     {
-        let (v, filtered_indices) = self
-            .iter()
-            .enumerate()
-            .filter_map(|(original_idx, val)| {
-                filter(val.clone()).map(|mapped| (mapped, original_idx))
-            })
-            .unzip();
-
-        let inner = self.subscribe();
-        let original_len = self.len();
-        let inner = FilterImpl { inner, filtered_indices, original_len };
-        let sub = FilterMap { inner, filter };
-
-        (v, sub)
+        FilterMap::new((*self).clone(), self.subscribe(), f)
     }
 }
